refactor(substrate-bundle): split test bench scenarios
